pyproject.toml
[project]
name = "discord-friender"
version = "0.1.0"
description = "discord mass friend request tool"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "click>=8.3.1",
    "discord-py-self>=2.1.0",
    "nopecha>=2.0.1",
    "python-dotenv>=1.2.1",
]
